requirements/base.txt
Django==2.2.6
djangorestframework==3.10.3
django-environ==0.4.5
django-extensions==2.2.3
django-json-field==0.5.7
django-model-utils==3.2.0

python-dateutil==2.8.0


# aliyun
aliyun-python-sdk-core==2.13.10